+++ spec/support/factories.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :user do
     email      'test@example.com'
-    identities { [TestIdentity.new(:uid => 'testuser')] }
+    identities { [TestIdentity.new(:uid => 'testuser', token: 'test')] }
   end
 end