+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Identities are destroy dependent of parent
 * Identities are always autosave
 * Identities will require validation for parent to save
+* Identities require presence of token